2019-05-25 by chisj

+++ src/main/java/org/study/heat/resolvers/CurrentUserMethodArgumentResolver.java
@@ -45,7 +45,7 @@ public class CurrentUserMethodArgumentResolver implements HandlerMethodArgumentR
     public Object resolveArgument(MethodParameter methodParameter, ModelAndViewContainer modelAndViewContainer,
                                   NativeWebRequest nativeWebRequest, WebDataBinderFactory webDataBinderFactory) throws Exception {
         //取出鉴权时存入的登录用户Id
-        Integer currentUserId = (Integer) nativeWebRequest.getAttribute(Constants.CURRENT_USER_ID, RequestAttributes.SCOPE_REQUEST);
+        Long currentUserId = (Long) nativeWebRequest.getAttribute(Constants.CURRENT_USER_ID, RequestAttributes.SCOPE_REQUEST);
         System.out.println("------------------------------chisj:currentUserId=" + currentUserId);
         if (currentUserId != null) {
             //从数据库中查询并返回
